requirements.txt
streamlit==1.39.0
langchain-community==0.3.4
langchain-openai==0.2.5
openai==v1.53.0
python-dotenv==1.0.1